VAE/requirements.txt
torch
torchvision
matplotlib
pyyaml
tensorboard
